+++ src/controllers/enrollmentReports/map/mapUtils/enrollment.ts
@@ -2,11 +2,13 @@ import { Child, Site, Enrollment } from '../../../../entity';
 import {
   AgeGroup,
   CareModel,
+  ChangeTag,
   ExitReason,
 } from '../../../../../client/src/shared/models';
 import { EnrollmentReportRow } from '../../../../template';
 import { mapEnum } from '.';
 import { getManager } from 'typeorm';
+import { MapResult } from '../uploadTypes';
 
 /**
  * Create Enrollment object from FlattenedEnrollment source,
@@ -80,4 +82,59 @@ export const getExitReason = (
   )
     return ExitReason.MovedWithinProgram;
   return ExitReason.Unknown;
+};
+
+/**
+ * Perform creation or update of a new enrollment for a child,
+ * if one is contained in the given spreadsheet row.
+ */
+export const handleEnrollmentUpdate = (
+  source: EnrollmentReportRow,
+  site: Site,
+  child: Child,
+  enrollmentsToUpdate: Enrollment[],
+  mapResult: MapResult,
+  matchingIdx: number
+) => {
+  const currentEnrollment: Enrollment | undefined = child.enrollments.find(
+    (e) => !e.exit
+  );
+
+  let enrollment = mapEnrollment(source, site, child);
+  const isNewEnrollment = rowHasNewEnrollment(
+    source,
+    currentEnrollment,
+    enrollment
+  );
+
+  if (isNewEnrollment) {
+    enrollment.fundings = [];
+    enrollmentsToUpdate.push(enrollment);
+    child.enrollments.push(enrollment);
+    // DESIGN NOTE: We'll guess that the previous enrollment and funding
+    // ended "one unit" (day/reporting period) before the new ones
+    currentEnrollment.exit = enrollment.entry.clone().add(-1, 'day');
+    currentEnrollment.exitReason = getExitReason(currentEnrollment, enrollment);
+    if (currentEnrollment.exitReason === ExitReason.AgedOut) {
+      mapResult.changeTagsForChildren[matchingIdx].push(ChangeTag.AgedUp);
+    } else {
+      mapResult.changeTagsForChildren[matchingIdx].push(
+        ChangeTag.ChangedEnrollment
+      );
+    }
+    enrollmentsToUpdate.push(currentEnrollment);
+  }
+
+  // If no new enrollment provided, new funding might apply
+  // to most current valid enrollment, so set that here so
+  // we don't have to think about it in map
+  else {
+    enrollment = currentEnrollment;
+  }
+
+  return {
+    enrollment,
+    currentEnrollment,
+    isNewEnrollment,
+  };
 };
